requirements.txt
# EYES-DEFY-ANEMIA requirements
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
openpyxl>=3.1.0          # read .xlsx files
Pillow>=9.4.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
streamlit>=1.28.0
plotly>=5.14.0